requirements.txt
fastapi
uvicorn
requests
python-dotenv
google-generativeai
youtube-transcript-api
duckduckgo-search
httpx
psutil
tenacity
